split fixtures run_upload into its own file
