+++ corehq/apps/fixtures/upload/__init__.py
@@ -1,9 +1,8 @@
+from corehq.apps.fixtures.upload.run_upload import run_upload, safe_fixture_upload
 from .upload import (
     DELETE_HEADER,
     FixtureWorkbook,
     get_workbook,
-    run_upload,
-    safe_fixture_upload,
     validate_file_format,
 )
 __all__ = [
